Addressed feedback
