+++ academy-ui/src/components/main/TopNav/TopNav.tsx
@@ -4,10 +4,10 @@ import { PredefinedSpaces } from '@/chatbot/utils/app/constants';
 import ButtonLarge from '@dodao/web-core/components/core/buttons/Button';
 import FullPageModal from '@dodao/web-core/components/core/modals/FullPageModal';
 import CreateContentModalContents from '@/components/main/TopNav/CreateContentModalContents';
-import { DesktopNavLink } from '@/components/main/TopNav/DesktopNavLink';
-import { DesktopProfileMenu } from '@/components/main/TopNav/DesktopProfileMenu';
-import { MobileNavLink } from '@/components/main/TopNav/MobileNavLink';
-import { MobileProfileMenu } from '@/components/main/TopNav/MobileProfileMenu';
+import { DesktopNavLink } from '@dodao/web-core/components/main/TopNav/DesktopNavLink';
+import { DesktopProfileMenu } from '@dodao/web-core/components/main/TopNav/DesktopProfileMenu';
+import { MobileNavLink } from '@dodao/web-core/components/main/TopNav/MobileNavLink';
+import { MobileProfileMenu } from '@dodao/web-core/components/main/TopNav/MobileProfileMenu';
 import { useLoginModalContext } from '@dodao/web-core/ui/contexts/LoginModalContext';
 import { Space, SpaceTypes, SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
 import { Session } from '@dodao/web-core/types/auth/Session';
