+++ base-ui/src/app/ChildLayout.module.scss
@@ -1,0 +1,5 @@
+.main {
+  background-color: var(--bg-color);
+  color: var(--text-color);
+  min-height: calc(100vh - 64px);
+}
