+++ base-ui/src/components/main/TopNav/TopNav.module.scss
@@ -1,0 +1,4 @@
+.main {
+  background-color: var(--bg-color);
+  color: var(--text-color);
+}
