+++ shared/web-core/src/components/main/TopNav/DesktopNavLink.module.scss
@@ -1,0 +1,8 @@
+.isActive {
+  border-bottom: 3px solid var(--primary-color);
+  color: var(--primary-color);
+}
+
+.isInactive {
+  color: var(--text-color);
+}
