+++ shared/web-core/src/components/main/TopNav/DesktopProfileMenu.module.scss
@@ -1,0 +1,4 @@
+.menuItem {
+  background-color: var(--bg-color);
+  border: 1px solid var(--border-color);
+}
