+++ shared/web-core/src/components/main/TopNav/MobileNavLink.module.scss
@@ -1,0 +1,4 @@
+.isActive {
+  border-bottom: 3px solid var(--primary-color);
+  color: var(--primary-color);
+}
